requirements.txt
# Core dependencies
neo4j>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.21.0
PyYAML>=6.0.0

# OpenAI integration (optional)
openai>=1.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Logging and utilities
python-dotenv>=1.0.0